56. Toast 로직 composition 함수로 추출
